refactor reporting job control route
